requirements.txt
anyio==4.9.0
asgiref==3.8.1
boto3==1.37.9
botocore==1.37.9
cachetools==5.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==4.2.16
django-cors-headers==4.7.0
django-drf-filepond==0.5.0
django-editorjs-fields==0.2.7
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.2
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
gspread==6.0.2
gspread_asyncio==2.0.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jmespath==1.0.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.11.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
StrEnum==0.4.15
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
whitenoise==6.9.0
celery==5.3.4
redis==5.0.1
